feat(kong-ee): adapt kong-build-tools to be more extendable refs: #ENGEN-472 (#452)
* feat(kong-ee): move enterprise specific build steps into kong-ee

* feat(kogn-ee): allow for pre and post build step hooks

* chore(ci): temporarily use a different Kong branch

* chore(ci): nuke the retry

* chore(kong-ee): add the openssl aux module logic

* fix(build): these steps happen after building openresty

* chore(packaging): theses ssh keys were never used

* chore(packaging): move the post build steps to later

* chore(packaging): add some convenience symlinks

* chore(Make): small makefile tweaks

* chore(ci): cleanup the Jenkinsfile

* chore(ci): add enterprise builds to the testing matrix

* fix(packaging): ee installs stuff here but ce does not

* chore(ci): cleanup ci stop building debian:8

* chore(ci): test kong-ee first

* chore(ci): recursive clone so it includes the submodules

* chore(ci): why can't I clone via ssh?

* fix(ci): does a ssh keyscan fix it?

* debug(ci): exactly who am I?

* chore(ci): try using a github key

* fix(ci): envrionment declared in the wrong spot?

* fix(ci): wait for apt to not be running

* chore(ci): re-use this workaround

* fix(c99): use c99 compat mode

* chore(backwards-compatible): keep us backwards compatible

* fix(backwards-compatible): adjustments so we're actually backwards compatible

* fix(cleanup): some adjustments to the submodules

* feat(make): set default target to "package-kong"

* fix(make): plumb GITHUB_TOKEN for openresty build

* fix(make): paths for lmdb luarock & nginx module

* test(enterprise): various enterprise only tests

* fix(jenkins): move GITHUB_TOKEN to kong build

* fix(jenkins): GITHUB_TOKEN jenkinsfile scope

* fix(docker): mv gh token ARG

* fix(docker): rm gh token from oresty dockerfile

* fix(ci): we require pulp credentials to pull the license file

* fix(ci): double check the past run is cleaned up

* fix(build): this file only exists sometimes

* chore(ci): more verbose log messages

* fix(tests): adjust how we cleanup between api and enterprise api tests

* fix(tests): this wait_for seems suspicious

* fix(ci): we require pulp credentials

* tests(portal): copy over the portal tests

* feat(make): announce which makefile we're using

* feat(admin/portal): relocate test + re-add rbac

* fix(ci): ensure perl/zlib1g installed for tests

* fix(make): (cleanup) cleanup-tests container

* fix(make): manage sub-modules

* chore(ci): freshen buildx/docker-machine

* fix(admin/portal): enterprise RBAC test

* feat(docker): accept DOCKER_BUILD_PROGRESS

* feat(bytecompile): optionally bytecompile

* fix(bytecompile): just run script if present

* fix(ci): typo in build-test-container target

* feat(mac): skip systemd pkg tests on Mac OS

* feat(ci): simplify Jenkinsfile make invocation

* chore(jenkins): mv DOCKER_REPOSITORY/KONG_SOURCE

* feat(packaging): adapt so kong-ee can package correctly

* chore(ci): adapt how we infer if Kong is alive

* fix(packaging): default to kong community edition when arguments are empty

* chore(make): print RESTY_IMAGE_{BASE,TAG}

* chore(cleanup): these variables are all bintray related

* fix(make): skip daemon.json write w/o /etc/docker

* feat(©): maybe use post-copyright-manifests.sh

* chore(ci): do the test alpine arm64 build

* fix(jenkins) re-introduce rockylinux 8

* Revert "chore(ci): do the test alpine arm64 build"

This reverts commit 3ecadace1f9e3dc443572aea5afb551a0ca11cbf.

Seems this hasn't worked for a while. Going to fix it out of band if at all

* chore(ci): leave CI test builds on master

* fix(ci): include the kong/distribution directory in the openresty cache key

* chore(release): need to set the docker repository we're releasing to but default to kong/kong

* Update build-kong.sh


* test(routes): include a list routes e2e test

* fix(enterprise): we need the enterprise flag when running the release scripts container

* chore(ci): align the test targets with what we release. Add a retry

+++ test/tests/01-package/run.sh
@@ -18,6 +18,7 @@ fi
 if [[ "$PACKAGE_TYPE" == "deb" ]]; then
   cp $PACKAGE_LOCATION/*amd64.deb kong.deb
   docker exec ${USE_TTY} user-validation-tests /bin/bash -c "apt-get update"
+  docker exec ${USE_TTY} user-validation-tests /bin/bash -c "apt-get install -y perl-base zlib1g-dev"
   docker exec ${USE_TTY} user-validation-tests /bin/bash -c "dpkg -i /src/kong.deb || apt install --fix-broken -y"
   docker exec ${USE_TTY} user-validation-tests /bin/bash -c "kong version"
 fi
@@ -107,11 +108,19 @@ docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "ls -l
 docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "ls -l /usr/local/kong/lib/pluginsocket.proto"
 docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "ls -l /usr/local/kong/lib/google/protobuf/*.proto"
 
+if [[ "$EDITION" == "enterprise" ]]; then
+  docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /usr/local/openresty/bin/resty -e 'require("ffi").load "passwdqc"'
+  docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /usr/local/openresty/bin/resty -e 'require("ffi").load "jq"'
+  #docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} openapi2kong 2>&1 | head -1 | grep 'missing required parameter:'
+  docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} ls -l /usr/local/kong/lib/pluginsocket.proto
+  docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/bash -c "ls -l /usr/local/kong/lib/google/protobuf/*.proto"
+fi
+
 # kong binaries
 docker run ${USE_TTY} --user=root --rm ${KONG_TEST_IMAGE_NAME} /bin/sh -c "/usr/local/kong/bin/openssl version | grep -q ${RESTY_OPENSSL_VERSION}"
 
 # TODO enable this test in other distros containing systemd
-if [[ "$RESTY_IMAGE_BASE" == "ubuntu" ]]; then
+if [[ "$RESTY_IMAGE_BASE" == "ubuntu" ]] && [ -z "${DARWIN:-}" ]; then
   cp $PACKAGE_LOCATION/*amd64.deb kong.deb
   docker run -d --rm --name systemd-ubuntu -e KONG_DATABASE=off --privileged -v /sys/fs/cgroup:/sys/fs/cgroup:ro -v $PWD:/src jrei/systemd-ubuntu:$RESTY_IMAGE_TAG
   docker exec ${USE_TTY} systemd-ubuntu /bin/bash -c "apt-get clean"
